test indexing bug + minor changes + dense distance

+++ cell2location/models/reference/_reference_module.py
@@ -176,8 +176,8 @@ class RegressionBackgroundDetectionTechPyroModel(PyroModule):
         }
 
     def forward(self, x_data, idx, batch_index, label_index, extra_categoricals):
-        obs2sample = one_hot(batch_index, self.n_batch)
-        obs2label = one_hot(label_index, self.n_factors)
+        obs2sample = one_hot(batch_index, self.n_batch).float()
+        obs2label = one_hot(label_index, self.n_factors).float()
         if self.n_extra_categoricals is not None:
             obs2extra_categoricals = torch.cat(
                 [
@@ -188,7 +188,7 @@ class RegressionBackgroundDetectionTechPyroModel(PyroModule):
                     for i, n_cat in enumerate(self.n_extra_categoricals)
                 ],
                 dim=1,
-            )
+            ).float()
 
         obs_plate = self.create_plates(x_data, idx, batch_index, label_index, extra_categoricals)
